refactor directory structure

+++ build.sh
@@ -1,18 +1,19 @@
 #!/usr/bin/env bash
-QF_VERSION=0.0.10
-
 WITH_MYSQL=${WITH_MYSQL:-0}
 
 arguments=(--with-python --with-openssl)
-if [ "$WITH_MYSQL" -eq 1 ]; then
-    arguments+=(--with-mysql)
+
+# TODO: support building with mysql and postgresql support
+# if [ "$WITH_MYSQL" -eq 1 ]; then
+#     arguments+=(--with-mysql)
+# fi
+
+if [ -d quickfix ]; then
+  echo "quickfix exists."
+else
+    git clone https://github.com/pablodcar/quickfix
+    rm -rf quickfix/.git
 fi
-
-rm -rf quickfix
-
-# git clone https://github.com/pablodcar/quickfix
-cp -r ../quickfix .
-rm -rf quickfix/.git
 
 # pushd quickfix/doc
 # ./document.sh
@@ -20,19 +21,11 @@ rm -rf quickfix/.git
 
 pushd quickfix || exit
 ./bootstrap
-popd || exit
-
-rm -f quickfix-$QF_VERSION.tar.gz
-
-tar czvf quickfix-$QF_VERSION.tar.gz quickfix
-
-pushd quickfix || exit
 ./configure "${arguments[@]}" && make
-# make check
+# make check TODO: tests need to be fixed in the original quickfix
 popd || exit
 
 pushd quickfix/src/python || exit
-# building Python interface
+echo "building Python interface..."
 ./swig.sh
 popd || exit
-
